[OPR-606]: Moves async block tests into same file
